+++ galasa-ui/src/components/headers/PageHeader.tsx
@@ -7,7 +7,9 @@
 
 import { Header, HeaderName, SkipToContent, Theme } from '@carbon/react';
 import PageHeaderMenu from "./PageHeaderMenu";
-import LeftHeaderMenu from './LeftHeaderMenu';
+import Image from 'next/image';
+import galasaLogo from "@/assets/images/galasaLogo.png";
+import Link from 'next/link';
 
 export default function PageHeader({ galasaServiceName }: { galasaServiceName: string }) {
 
@@ -17,11 +19,18 @@ export default function PageHeader({ galasaServiceName }: { galasaServiceName: s
 
         <SkipToContent />
 
-        <LeftHeaderMenu />
+        <Link href={"/"} style={{"paddingLeft" : "0.5rem"}}>
+          <Image
+            src={galasaLogo}
+            width={28}
+            height={28}
+            alt='Galasa logo'
+          />
+        </Link>
 
-        <HeaderName prefix="">{galasaServiceName}</HeaderName>
+        <HeaderName href="/" prefix="">Galasa</HeaderName>
 
-        <PageHeaderMenu />
+        <PageHeaderMenu galasaServiceName={galasaServiceName} />
 
       </Header>
     </Theme>
